requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
streamlit